broke out into individual test.py
